another batch of inspection related code changes

+++ src/main/java/xreliquary/items/block/ItemFertileLilyPad.java
@@ -57,7 +57,7 @@ public class ItemFertileLilyPad extends ItemBlockBase {
 			BlockPos blockpos1 = blockpos.up();
 			IBlockState iblockstate = worldIn.getBlockState(blockpos);
 
-			if(iblockstate.getBlock().getMaterial(iblockstate) == Material.WATER && ((Integer) iblockstate.getValue(BlockLiquid.LEVEL)).intValue() == 0 && worldIn.isAirBlock(blockpos1)) {
+			if(iblockstate.getBlock().getMaterial(iblockstate) == Material.WATER && iblockstate.getValue(BlockLiquid.LEVEL) == 0 && worldIn.isAirBlock(blockpos1)) {
 				// special case for handling block placement with water lilies
 				net.minecraftforge.common.util.BlockSnapshot blocksnapshot = net.minecraftforge.common.util.BlockSnapshot.getBlockSnapshot(worldIn, blockpos1);
 				worldIn.setBlockState(blockpos1, ModBlocks.fertileLilypad.getDefaultState());
